UPSTREAM: <carry>: e2e add Dockefiles for OpenShift CI builds
We want to run the e2e suite, but we want to run it against something
like what we're actually going to release, so instead of using the
synthetic setup/teardown functions from the keda tests, we try to pack a
proper bundle and install from that.

This contains two dockerfiles. One for the operator which:
- Builds the operator with the openshift builder/release images instead
  of the upstream builder image + distroless
And another for the bundle which:
- Attempts to perform the "release steps" against the repo to generate a
  new v0.0.0 bundle that can be built by OpenShift CI's bundle builder,
  and will allow for image overrides