UPSTREAM: <carry>: e2e add Dockefiles for OpenShift CI builds
We want to run the e2e suite, but we want to run it against something
like what we're actually going to release, so instead of using the
synthetic setup/teardown functions from the keda tests, we try to pack a
proper bundle and install from that.

This contains two dockerfiles. One for the operator which:
- Builds the operator with the openshift builder/release images instead
  of the upstream builder image + distroless
And another for the bundle which:
- Can be used by the openshift CI bundle builder to build a cma bundle
  approximating that which is built by osbs

Signed-off-by: John Kyros <jkyros@redhat.com>