Update audit configmap when config changes
Previously if the configmap for the audit policy already existed, and
the audit policy in the kedacontroller was updated, the keda controller
would not update the configmap, resulting in the audit policy being
"stuck" at whatever was initially configured.

This adjusts the keda controller such that it generates the audit config
map data regardless, compares what it generated to what's present in the
cluster, and then updates the configmap if they differ.

Signed-off-by: John Kyros <jkyros@redhat.com>