Make loglevel test actually wait for changes
Previously the functionality test that tests the log level changes was
not waiting for rollouts to happen, so it was effectively not testing
what it should have been, and we were getting lucky on timing.

Recent releases have changed that timing so this test has started to
fail.

This changes the test such that it now waits for controller to process
each of the changes (or not process each of the changes, depending on
the test case). It does this by adding an additional deployment
annotation for each test case, so the test case knows that the
controller has processed the test case and it can check the values.

This probably needs a more complete refactor at some point, but this at
least makes the test dependable.

Signed-off-by: John Kyros <jkyros@redhat.com>