UPSTREAM: <carry>: Set root fs to read-only in CMA 2.14.1 csv
This just sets readOnlyRootFilesystem: true in the pod spec for our
operator in the CMA 2.14.1 CSV because we rebased and generated it
before we had the fix.

This can get squashed into the CSV updates next time we rebase, but we
need to keep this one -- we generate CMA the CSV off the upstream CSV,
which means a rebase will not update the one we already generated, so
we need to carry this forever.

Signed-off-by: John Kyros <jkyros@redhat.com>